Add commands to generate config, deploy token
Add command to generate a config file.

This config file should contain the right values for RPC urls so that it
can be used in production. It generates a new mnemonic so we have less
to document about how to do that. The docs will just have to tell the
requests funds from the faucet(s) for this address.

Add a command to deploy a token. This makes deployment and documentation
easier because users can easily deploy their own tokens to experiment
with and don't need us to provide them.

Apologies there are also a bunch of miscellaneous changes that I ran
into while working on this.

Try not to show tracebacks to users (unless there are internal
errors) and instead show a simple and clear error message and exit with
non-zero exit code.

Support a few more defaults when minting: mint to the owners address by
default, mint 10000 tokens by default.

Set polling interval to 2 seconds, 7 seconds makes sense for Ethereum
mainnet but for our sequencer it is on the slow side because that window
may contain several blocks.

Leave commands out of config file. I think it doesn't really make sense
to have them in there because users are expected to switch between
commands often.