Increase funded timeout
I need a longer timeout for this to pass locally. I guess the
important thing is that it passes in CI but it seems that a common
debugging scenario would be to run this locally, and it causes
confusion if it fails. Also if this succeeds before the timeout it
continues anyway so there is not much of a downside.